Project 2 - Error Handling Overhaul
The current thread's last_error is now initialized to ENONE everytime the kernel is entered.  This eliminates the need for kernel functions to explicitly set ENONE (except for kthread_create which is called from within the kernel to spawn the idle and main threads).

errno is now used in the kernel rather than checking current_thread for NULL and then setting current_thread->last_error.