Project 2 - Preemption Fix
Previously the timer ISR which preempted threads was using a syscall (SYSCALL_YIELD) to give control of the CPU back to the operating system.  However there was the possibility that the interrupt occurred while the current thread was initiating a syscall, thereby corrupting the syscall.

Gave the timer ISR a special path into the kernel to rectify this.