Project 2 - Context Switching Simplification
The macros SAVE_CTX and RESTORE_CTX save/restore all registers (including RAMPZ and EIND) as well as the status register.  However these macros are embedded in functions (kenter_impl and kexit_impl) and the compiler already saves/restores registers as needed when calling/returning from a function.  Therefore the context we are interested in saving/restoring is already saved on the stack, and therefore all we have to do is switch out the stack pointers.  Saving/restoring SREG is also important due to the global interrupt enable bit.

Accordingly eliminated the vast majority of pushes/pops in SAVE_CTX and RESTORE_CTX.  Everything appears to be working just fine in their absence.  Context switching overhead is much lower.