Project 2 - Test 4
When compiling with the preprocessor macro TEST equal to 4 the fourth provided test is now built.

Had to modify kernel.c to release timer 3 (the sleep timer) when building the fourth provided test since it makes use of this timer (would've been nice to know that before I built the sleep infrastructure around the assumption that user mode tasks wouldn't use the hardware timers).